Allow client to ensure data has not been tampered
Question: How does the client insure that their data has not been
tampered with? Assume that the database is compromised.

Solution: The data is signed via sha256 checksum, though this may be
cryptographically signed as well. In any case, the client saves a copy
of this checksum in localStorage, and the "Verify" checks the data
against this checksum.