Allow client to recover from tampered data
Question: If the data has been tampered with, how can the client recover
the lost data?

Solution: Client data is saved to localStorage when the client updates
the data. This is because we know that only the user is able to update
their own data, and so we are able to ensure both the checksum and data
are correct.

When the user attempts to verify data and is notified that the data has
been tampered on the server, the user is shown a button to recover their
data from localStorage, which will then update the data on the server.