Update README.md
Dear Bequest Team,

I would like to express my gratitude for the opportunity to respond to the questions related to data security in my application. I am excited about the possibility of contributing to the team and further enhancing data integrity standards at Bequest.

Feel free to reach out if there is a need for additional clarifications or supplementary information. I look forward to the possibility of working with such a respected and dedicated team.

Thank you again for considering my application.

Best regards,
Wagner Sanches